+++ CHANGELOG.md
@@ -18,6 +18,7 @@
   - key `file_type` was removed from endpoints GET [Publications](doc/rest.md#get-publications)/[Layers](doc/rest.md#get-layers)/[Workspace Layers](doc/rest.md#get-workspace-layers)/[Workspace Layer](doc/rest.md#get-workspace-layer) response
   - key `file`.`path` was removed from GET [Workspace Layer](doc/rest.md#get-workspace-layer) response
   - key `sld` was removed from GET [Workspace Layer](doc/rest.md#get-workspace-layer) response
+  - key `db_table` was removed from GET [Workspace Layer](doc/rest.md#get-workspace-layer) response
   - body parameter `sld` was removed from [POST Workspace Publications](doc/rest.md#post-workspace-layers) and [PATCH Workspace Publication](doc/rest.md#patch-workspace-layer)
 - [#1028](https://github.com/LayerManager/layman/issues/1028) Upgrade Node.js of Laymen Test Client from v18 to v22 and dependencies:
   - eslint-config-next 13 -> 14